add permission check when get private cred

+++ services/cert_manager_standard/cert_manager_engine/main/core/src/cert_manager_file.c
@@ -83,7 +83,7 @@ static int32_t MallocFileNames(struct CmMutableBlob *fileNames, const char *path
         return -1;
     }
 
-    for (uint32_t i = 0; i < fileNums; i++) {
+    for (uint32_t i = 0; i < (uint32_t)fileNums; i++) {
         tmp[i].data = NULL;
         tmp[i].size = 0;
     }
